# awesomeProject/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for User, Book Storage.\nWhere they can create, retrieve, update, delete books.\nAnd can rent these books",
        "title": "OneLab HomeWork",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/books": {
            "post": {
                "description": "create book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Create",
                "operationId": "BookCreate",
                "parameters": [
                    {
                        "description": "book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/books/:id": {
            "delete": {
                "description": "Delete book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "Delete",
                "operationId": "BookDelete",
                "parameters": [
                    {
                        "description": "book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/books/:is": {
            "get": {
                "description": "get book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BookGet"
                ],
                "summary": "Get",
                "operationId": "Book",
                "parameters": [
                    {
                        "description": "book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/cards": {
            "post": {
                "description": "create card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Create",
                "operationId": "Card",
                "parameters": [
                    {
                        "description": "card info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/rent/:id": {
            "post": {
                "description": "make request to another service to create record about renting of book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History of Book"
                ],
                "summary": "Rent",
                "operationId": "Rent",
                "parameters": [
                    {
                        "description": "history book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HistoryBook"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/rent/debtors": {
            "get": {
                "description": "make request to another service to get records of users with book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History of Book"
                ],
                "summary": "Debtors",
                "operationId": "Debtors",
                "parameters": [
                    {
                        "description": "history book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HistoryBook"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/rent/incomes": {
            "get": {
                "description": "make request to another service to get list of incomes of each books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History of Book"
                ],
                "summary": "Incomes",
                "operationId": "Income",
                "parameters": [
                    {
                        "description": "history book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Income"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/rent/return/:id": {
            "post": {
                "description": "make request to another service to update record when user return book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History of Book"
                ],
                "summary": "Return",
                "operationId": "Return",
                "parameters": [
                    {
                        "description": "history book info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/logout": {
            "delete": {
                "description": "delete session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Logout",
                "operationId": "delete-session",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/signin": {
            "post": {
                "description": "authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "SignIn",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/signup": {
            "post": {
                "description": "create account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sign-Up",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.Card": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "money": {
                    "type": "number"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.HistoryBook": {
            "type": "object",
            "properties": {
                "bookID": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isGiven": {
                    "type": "boolean"
                },
                "isPaid": {
                    "type": "boolean"
                },
                "price": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Income": {
            "type": "object",
            "properties": {
                "bookID": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}